Add property componet inside properties container; refactor code
