+++ src/containers/properties/Property.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Property = () => {
+    return (
+        <div>
+            <h1>Property</h1>
+        </div>
+    )
+}
+
+export default Property
